Crud finalizado com produtos

+++ DigitalMenu/src/main/java/com/br/digitalmenu/repository/DiaSemanaRepository.java
@@ -3,5 +3,5 @@ package com.br.digitalmenu.repository;
 import com.br.digitalmenu.model.DiaSemana;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaSemanaRepository extends JpaRepository<DiaSemana,Integer> {
+public interface DiaSemanaRepository extends JpaRepository<DiaSemana,Long> {
 }
