+++ DigitalMenu/src/main/java/com/br/digitalmenu/repository/RestricaoRepository.java
@@ -3,5 +3,5 @@ package com.br.digitalmenu.repository;
 import com.br.digitalmenu.model.Restricao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestricaoRepository extends JpaRepository<Restricao,Integer> {
+public interface RestricaoRepository extends JpaRepository<Restricao,Long> {
 }
